requirements.txt
cdsapi~=0.5.1
requests~=2.28.1
pandas~=1.4.3
xarray~=0.20.1
beautifulsoup4~=4.11.1
rioxarray~=0.12.0
rasterio~=1.3.2
numpy~=1.23.2
matplotlib~=3.5.3
earthaccess~=0.8.1
setuptools~=63.4.1
geopandas~=0.11.1
wget~=3.2
exactextract~=0.2.0.dev0
python-dotenv~=1.0.0